refactor(warehouse-core): promueve el VO Capacity de logistics al kernel
Mueve Capacity + CapacityProps y sus errores (CapacityMustHaveWeightOrVolumeError,
InvalidCapacityAmountError) al kernel, para que inventory y logistics lo compartan
sin dependencia cruzada. logistics/capacity.ts y transport-capacity-errors.ts pasan
a re-exportar desde el kernel, sin romper a sus consumidores.
